requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1

# Database (SQLite by default for local run; PostgreSQL for Docker/production)
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4

# Validation and settings
pydantic[email]==2.6.1
pydantic-settings==2.1.0

# Redis
redis==5.0.1

# Celery
celery[redis]==5.3.6

# Utils
python-multipart==0.0.9
python-dotenv==1.0.1
